Validate `topicId` and `Topic.GetUniqueKey()` before deleting
Previously, there was a bug where an attempt to delete a draft Topic—i.e., a topic that is saved in memory, but hasn't yet been committed to the database—would result in the code attempting to delete the _entirely_ of the database. This is because a draft topic has a temporary `Id` of `-1`, but `-1` is used by `TopicRepository.Load()` as a special instruction to load the entire topic graph. In practice, this doesn't result in the deletion of the entire topic graph because the `DeleteTopic` stored procedure is optimized for deleting a small number of topics, and will timeout in this scenario. Nevertheless, the delete fails and, worse, the database gets jammed up for a bit before it times out, thus resulting in other database calls from potentially failing. This scenario should never occur, as there shouldn't be any draft topics in these locations. But if there's an unexpected connection error or downtime with the database, it becomes a possibility. To prevent this scenario from occuring, this measure validates that the `topicId` is a positive number, and additionally validates that the resulting topic is, in fact, within the expected tree. If it's not, the record is skipped.

+++ Areas/Administration/Controllers/InvoicesController.cs
@@ -172,7 +172,13 @@ namespace GoldSim.Web.Administration.Controllers {
       | Delete topics
       \-----------------------------------------------------------------------------------------------------------------------*/
       foreach (var topicId in topics) {
+        if (topicId < 0) {
+          continue;
+        }
         var topic = _topicRepository.Load(topicId);
+        if (!topic.GetUniqueKey().StartsWith(_invoiceRoot, StringComparison.InvariantCultureIgnoreCase)) {
+          continue;
+        }
         _topicRepository.Delete(topic);
       }
